+++ manager/internal/manager/update_installer_other.go
@@ -1,0 +1,7 @@
+//go:build !windows
+
+package manager
+
+func newPlatformUpdateInstaller(_ string) updateInstallController {
+	return disabledUpdateInstaller{}
+}